+++ src/integrationTest/kotlin/org/jetbrains/intellij/platform/gradle/MultiModuleIntegrationTest.kt
@@ -37,7 +37,7 @@ class MultiModuleIntegrationTest : IntelliJPlatformIntegrationTestBase(
     @Test
     fun `custom task should refer to the composed jar of base module`() {
         val ext = dir.resolve("ext")
-        val sandboxDirectory = ext.resolve("build/idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+        val sandboxDirectory = ext.resolve(Constants.CACHE_DIRECTORY).resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
         val suffix = "_$CUSTOM_RUN_IDE_TASK_NAME"
         val pluginsDirectory = sandboxDirectory.resolve(Sandbox.PLUGINS + suffix)