+++ src/integrationTest/kotlin/org/jetbrains/intellij/platform/gradle/SandboxIntegrationTest.kt
@@ -18,7 +18,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
     @Test
     fun `create sandbox in a default location`() {
         build(Tasks.RUN_IDE, projectProperties = defaultProjectProperties) {
-            val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+            val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
 
             assertExists(sandboxDirectory)
             assertExists(sandboxDirectory.resolve(Sandbox.CONFIG))
@@ -31,7 +31,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
     @Test
     fun `create sandbox for a custom task in a default location`() {
         build(CUSTOM_RUN_IDE_TASK_NAME, projectProperties = defaultProjectProperties) {
-            val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+            val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
             val suffix = "_$CUSTOM_RUN_IDE_TASK_NAME"
 
             assertExists(sandboxDirectory)
@@ -45,7 +45,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
     @Test
     fun `create test sandbox in a default location`() {
         build(Tasks.External.TEST, projectProperties = defaultProjectProperties) {
-            val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+            val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
             val suffix = "-test"
 
             assertExists(sandboxDirectory)
@@ -59,7 +59,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
     @Test
     fun `create test sandbox for a custom task in a default location`() {
         build(CUSTOM_TEST_IDE_TASK_NAME, projectProperties = defaultProjectProperties) {
-            val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+            val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
             val suffix = "_$CUSTOM_TEST_IDE_TASK_NAME"
 
             assertExists(sandboxDirectory)
@@ -102,7 +102,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
                 """.trimIndent()
 
         build(Tasks.RUN_IDE, projectProperties = defaultProjectProperties) {
-            val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("custom-directory")
+            val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("custom-directory")
 
             assertExists(sandboxDirectory)
             assertExists(sandboxDirectory.resolve(Sandbox.CONFIG))
@@ -114,7 +114,7 @@ class SandboxIntegrationTest : IntelliJPlatformIntegrationTestBase(
 
     @Test
     fun `invalidate sandbox content when disabledPlugins property is changed`() {
-        val sandboxDirectory = buildDirectory.resolve("idea-sandbox").resolve("$intellijPlatformType-$intellijPlatformVersion")
+        val sandboxDirectory = buildDirectory.resolve(Sandbox.CONTAINER).resolve("$intellijPlatformType-$intellijPlatformVersion")
         val suffix = "_$CUSTOM_RUN_IDE_TASK_NAME"
         val disabledPluginsFile = sandboxDirectory.resolve(Sandbox.CONFIG + suffix).resolve("disabled_plugins.txt")